requirements.txt
aiogram==2.9.2
numpy==1.18.3
Pillow==7.1.2
dominate>=2.3.1
visdom>=0.1.8.3
urllib3>=1.25.9
http://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp38-cp38-linux_x86_64.whl
http://download.pytorch.org/whl/cpu/torchvision-0.6.1%2Bcpu-cp38-cp38-linux_x86_64.whl